+++ src/main/java/com/pt/pedrovcruzeiro/phonebook/PhonebookApplication.java
@@ -2,12 +2,16 @@ package com.pt.pedrovcruzeiro.phonebook;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cache.annotation.EnableCaching;
+import org.springframework.cloud.openfeign.EnableFeignClients;
 
 @SpringBootApplication
+@EnableFeignClients
+@EnableCaching
 public class PhonebookApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(PhonebookApplication.class, args);
-	}
+    public static void main(String[] args) {
+        SpringApplication.run(PhonebookApplication.class, args);
+    }
 
 }