+++ src/main/java/com/pt/pedrovcruzeiro/phonebook/configuration/WebMvcConfiguration.java
@@ -1,0 +1,7 @@
+package com.pt.pedrovcruzeiro.phonebook.configuration;
+
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class WebMvcConfiguration {
+}