+++ src/main/java/com/pt/pedrovcruzeiro/phonebook/controller/PhonebookControllerAdvice.java
@@ -1,0 +1,4 @@
+package com.pt.pedrovcruzeiro.phonebook.controller;
+
+public class PhonebookControllerAdvice {
+}