+++ src/main/java/com/pt/pedrovcruzeiro/phonebook/entity/dto/in/PhoneNumberBusinessSectorResponse.java
@@ -1,0 +1,14 @@
+package com.pt.pedrovcruzeiro.phonebook.entity.dto.in;
+
+import lombok.*;
+
+@Builder
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class PhoneNumberBusinessSectorResponse {
+
+    private String number;
+    private String sector;
+}