+++ src/main/java/com/pt/pedrovcruzeiro/phonebook/util/formatter/PhonebookLogFormatter.java
@@ -1,0 +1,24 @@
+package com.pt.pedrovcruzeiro.phonebook.util.formatter;
+
+import com.pt.pedrovcruzeiro.phonebook.util.constant.PhonebookConstants;
+import org.slf4j.MDC;
+
+import java.sql.Timestamp;
+import java.time.Instant;
+
+public class PhonebookLogFormatter {
+
+    private PhonebookLogFormatter(){
+
+    }
+
+    public static String format(String message){
+        return String.format(
+                PhonebookConstants.PHONEBOOK_PREFIX_MSG,
+                Timestamp.from(Instant.now()),
+                MDC.get(PhonebookConstants.API_OPERATION),
+                message
+        );
+    }
+
+}