+++ src/test/java/com/pt/pedrovcruzeiro/phonebook/controller/PhoneBookControllerImplTest.java
@@ -1,0 +1,5 @@
+package com.pt.pedrovcruzeiro.phonebook.controller;
+
+class PhoneBookControllerImplTest {
+
+}