+++ CHANGELOG.md
@@ -1,13 +1,23 @@
 # Changelog
+
 All technical changes ill be documented in this file.
 
 ## [Unreleased]
 
-## [0.0.1] - 2020-10-03
+## [0.0.1] - 2021-12-08
+
 ### Added
+
 * Home endpoint.
+* Added Aggregation operation for Phonebook API
+* Added validation of the phone numbers
+
 ### Changed
+
 ### Deprecated
+
 ### Removed
+
 ### Fixed
+
 ### Security